Move remaining torii logic
